refactor: move headers to include/AesiMultiprecision/, integrate CMake, add install rules
Headers:
- Aeu.h and Aesi.h moved to include/AesiMultiprecision/
- All #include directives updated to <AesiMultiprecision/Aeu.h> / <AesiMultiprecision/Aesi.h>

CMake:
- Root CMakeLists.txt declares INTERFACE library with proper BUILD/INSTALL include paths
- Added AESI_BUILD_TESTS, AESI_BUILD_BENCHMARKS, AESI_BUILD_SANITIZERS options
- test/, benchmark/, sanitize/unsigned/ now addable from root via those flags
- ci/build and ci/integration link against AesiMultiprecision target instead of
  manually setting include paths — include dir propagates via INTERFACE automatically

Install:
- install(DIRECTORY include/) exports headers to CMAKE_INSTALL_INCLUDEDIR
- Exports AesiMultiprecisionTargets.cmake for find_package support
- Generates AesiMultiprecisionConfig.cmake and AesiMultiprecisionConfigVersion.cmake
- cmake/AesiMultiprecisionConfig.cmake.in added

Release workflow:
- Updated to zip include/ directory instead of root-level header files
