+++ .github/workflows/upload_release.yml
@@ -20,7 +20,7 @@ jobs:
 
       - name: Generate release archive
         run: |
-          zip Aesi_${{ inputs.version || github.ref_name }}.zip Aesi.h Aeu.h README.md
+          zip -r Aesi_${{ inputs.version || github.ref_name }}.zip include/ README.md
 
       - name: Upload artifacts
         uses: actions/upload-artifact@v4
