+++ test/arithmetic/square-root.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
-#include "../../Aeu.h"
-#include "../../Aesi.h"
+#include <AesiMultiprecision/Aeu.h>
+#include <AesiMultiprecision/Aesi.h>
 #include "../generation.h"
 
 TEST(Signed_SquareRoot, SquareRoot) {
