+++ test/operations/initialization/unsigned.cpp
@@ -1,6 +1,6 @@
 #include <gtest/gtest.h>
 #include <bitset>
-#include "../../../Aeu.h"
+#include <AesiMultiprecision/Aeu.h>
 #include "../../generation.h"
 
 TEST(Unsigned_Initialization, Basic) {
